feat: refactor schedule management
